Address comments partially
